Move `make_data_buffer`.
This factory method should reside in the same file as DataBuffer.

Creating a DataBuffer is also required from other parts of the project,
for example for storing multiple status or request objects.
