+++ tests/data_buffer_test.cpp
@@ -1,6 +1,6 @@
 // This file is part of KaMPIng.
 //
-// Copyright 2021-2023 The KaMPIng Authors
+// Copyright 2021-2024 The KaMPIng Authors
 //
 // KaMPIng is free software : you can redistribute it and/or modify it under the terms of the GNU Lesser General Public
 // License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
@@ -789,3 +789,327 @@ TEST(LibAllocatedContainerBasedBufferTest, prevent_usage_after_extraction_via_mp
     EXPECT_KASSERT_FAILS(result.extract_send_recv_type(), "Cannot extract a buffer that has already been extracted.");
 }
 #endif
+
+TEST(DataBufferTest, make_data_buffer) {
+    {
+        // Constant, container, referencing, user allocated
+        std::vector<int>                  vec;
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(vec);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&vec, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<int> const&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, container, referencing, user allocated
+        std::vector<int>                  vec;
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::modifiable, btype, BufferResizePolicy::grow_only>(vec
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::grow_only);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&vec, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<int>&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, single element, referencing, user allocated
+        int                               single_int;
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(
+                single_int
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_TRUE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&single_int, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, int const&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, container, owning, user allocated
+        std::vector<int>                  vec;
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(
+                std::move(vec)
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<int> const>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+
+    {
+        // modifiable, container, owning, library allocated
+        constexpr internal::ParameterType ptype    = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype    = internal::BufferType::in_buffer;
+        auto                              data_buf = internal::make_data_buffer<
+            ptype,
+            BufferModifiability::modifiable,
+            btype,
+            BufferResizePolicy::grow_only>(alloc_new<std::vector<int>>);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::grow_only);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<int>>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, single element, owning, lib_allocated
+        constexpr internal::ParameterType ptype    = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype    = internal::BufferType::in_buffer;
+        auto                              data_buf = internal::
+            make_data_buffer<ptype, BufferModifiability::modifiable, btype, BufferResizePolicy::no_resize>(alloc_new<
+                                                                                                           int>);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_TRUE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, int>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, container, owning, user_allocated with initializer_list
+        constexpr internal::ParameterType ptype         = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype         = internal::BufferType::in_buffer;
+        constexpr BufferResizePolicy      resize_policy = BufferResizePolicy::no_resize;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::modifiable, btype, resize_policy>({1, 2, 3});
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, resize_policy);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<int>>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, container, owning, user_allocated with initializer_list
+        constexpr internal::ParameterType ptype         = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype         = internal::BufferType::in_buffer;
+        constexpr BufferResizePolicy      resize_policy = BufferResizePolicy::no_resize;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, resize_policy>({1, 2, 3});
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, resize_policy);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, const std::vector<int>>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+}
+
+TEST(DataBufferTest, make_data_buffer_boolean_value) {
+    // use a custom container, because std::vector<bool> is not supported (see compilation failure tests)
+    {
+        // Constant, container, referencing, user allocated
+        ::testing::OwnContainer<bool>     vec           = {true, false};
+        constexpr internal::ParameterType ptype         = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype         = internal::BufferType::in_buffer;
+        constexpr BufferResizePolicy      resize_policy = BufferResizePolicy::no_resize;
+        auto data_buf = internal::make_data_buffer<ptype, BufferModifiability::constant, btype, resize_policy>(vec);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, resize_policy);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&vec, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, ::testing::OwnContainer<bool> const&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, container, referencing, user allocated
+        ::testing::OwnContainer<bool>     vec      = {true, false};
+        constexpr internal::ParameterType ptype    = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype    = internal::BufferType::in_buffer;
+        auto                              data_buf = internal::
+            make_data_buffer<ptype, BufferModifiability::modifiable, btype, BufferResizePolicy::resize_to_fit>(vec);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::resize_to_fit);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&vec, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, ::testing::OwnContainer<bool>&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, single element, referencing, user allocated
+        bool                              single_bool;
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(
+                single_bool
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_TRUE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        // As this buffer is referencing, the addresses of vec ad data_buf.underlying() should be the same.
+        EXPECT_EQ(&single_bool, &data_buf.underlying());
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, bool const&>,
+            "Referencing buffers must hold a reference to their data."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_FALSE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, container, owning, user allocated
+        ::testing::OwnContainer<bool>     vec   = {true, false};
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(
+                std::move(vec)
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, ::testing::OwnContainer<bool> const>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+
+    {
+        // modifiable, container, owning, library allocated
+        constexpr internal::ParameterType ptype    = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype    = internal::BufferType::in_buffer;
+        auto                              data_buf = internal::make_data_buffer<
+            ptype,
+            BufferModifiability::modifiable,
+            btype,
+            BufferResizePolicy::resize_to_fit>(alloc_new<::testing::OwnContainer<bool>>);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::resize_to_fit);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, ::testing::OwnContainer<bool>>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, single element, owning, lib_allocated
+        constexpr internal::ParameterType ptype    = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype    = internal::BufferType::in_buffer;
+        auto                              data_buf = internal::
+            make_data_buffer<ptype, BufferModifiability::modifiable, btype, BufferResizePolicy::no_resize>(alloc_new<
+                                                                                                           bool>);
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_TRUE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, bool>,
+            "Owning buffers must hold their data directly."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Modifiable, container, owning, user_allocated with initializer_list
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::modifiable, btype, BufferResizePolicy::no_resize>(
+                {true, false, true}
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_TRUE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, std::vector<kabool>>,
+            "Initializer lists of type bool have to be converted to std::vector<kabool>."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+    {
+        // Constant, container, owning, user_allocated with initializer_list
+        constexpr internal::ParameterType ptype = internal::ParameterType::send_buf;
+        constexpr internal::BufferType    btype = internal::BufferType::in_buffer;
+        auto                              data_buf =
+            internal::make_data_buffer<ptype, BufferModifiability::constant, btype, BufferResizePolicy::no_resize>(
+                {true, false, true}
+            );
+        EXPECT_EQ(data_buf.parameter_type, ptype);
+        EXPECT_FALSE(data_buf.is_modifiable);
+        EXPECT_FALSE(data_buf.is_single_element);
+        EXPECT_EQ(data_buf.resize_policy, BufferResizePolicy::no_resize);
+        static_assert(
+            std::is_same_v<decltype(data_buf)::MemberTypeWithConstAndRef, const std::vector<kabool>>,
+            "Initializer lists of type bool have to be converted to std::vector<kabool>."
+        );
+        // extract() as proxy for lib allocated DataBuffers
+        EXPECT_TRUE(has_extract_v<decltype(data_buf)>);
+    }
+}
